requirements.txt
ipython
pygments
pdbpp
ipdb
isort
black<25

Django==6.0
django-stubs==5.2.8
django-extensions==4.1
django-dbbackup==5.1.0

djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0

pydantic==2.12.5
gunicorn==23.0.0
psycopg2-binary==2.9.11

PyMuPDF==1.26.7
spacy==3.8.11

